+++ lib/policy_machine.rb
@@ -292,6 +292,14 @@ class PolicyMachine
     end
   end
 
+  def all_operations_for_user_or_attr_and_objs_or_attrs(user_or_attribute, object_attribute_ids, options = {})
+    unless policy_machine_storage_adapter.respond_to?(:all_operations_for_user_or_attr_and_objs_or_attrs)
+      adapter = policy_machine_storage_adapter.class
+      raise NoMethodError, "all_operations_for_user_or_attr_and_objs_or_attrs is not implemented for storage adapter #{adapter}"
+    end
+    policy_machine_storage_adapter.all_operations_for_user_or_attr_and_objs_or_attrs(user_or_attribute, object_attribute_ids, options)
+  end
+
   def accessible_objects_for_operations(user_or_attribute, operations, options = {})
     unless policy_machine_storage_adapter.respond_to?(:accessible_objects_for_operations)
       adapter = policy_machine_storage_adapter.class